% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_histories.R
\name{add_histories}
\alias{add_histories}
\title{Add Specified Histories of Specified Columns to Data Set}
\usage{
add_histories(D, variables, histories)
}
\arguments{
\item{D}{data set}

\item{variables}{a vector containing variable names in the data set}

\item{histories}{a vector containing elements of the type: "lag_k", "cumsum", "cummean", "sum_prev_k"}
}
\value{
data frame
}
\description{
The j'th entry in histories is the desired history to be added from the j'th entry in variables
}
\examples{
add_histories(gfoRmula::basicdata[1:20,], c("L1","L2"), c("lag_1", "cumsum"))
add_histories(gfoRmula::basicdata[1:20,], c("L1","L2","L3","A"), c("lag_3", "cumsum", "cummean", "sum_prev_2"))
}
