% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier_CV.R
\name{brier_CV}
\alias{brier_CV}
\title{Brier-score based cross-validation of supplied p and q models}
\usage{
brier_CV(data, formulae, types)
}
\arguments{
\item{data}{data frame}

\item{formulae}{a vector of model formulations}

\item{types}{a vector of type of algorithm (glm or gam)}
}
\description{
The i'th element in formulae is the model formulation for the i'th
model type (algorithm) in types.
}
\examples{
brier_CV(add_histories(gfoRmula::basicdata, c("A","L1","L2"), c("lag_1","lag_1","lag_1")), 
        c("Y ~ A + L1 + L2 + L3 + A_lag_1 + L1_lag_1 + L2_lag_1 + t0", "Y ~ A + L1 + s(L2,df=4) + L3 + A_lag_1 + L1_lag_1 + L2_lag_1 + t0"),
        c("glm","gam"))
}
