% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ten_folds_indicator.R
\name{ten_folds_indicator}
\alias{ten_folds_indicator}
\title{Adds 10 fold indicators while respecting id's.}
\usage{
ten_folds_indicator(D)
}
\arguments{
\item{D}{data frame}
}
\value{
D with an addition column for the 10 fold indicator
}
\description{
Adds 10 fold indicators while respecting id's.
}
\examples{
ten_folds_indicator(gfoRmula::basicdata)
}
